refactor: 积分存储从厘分(INTEGER×10)改为分(REAL直接存)
【动机】 去掉全部 *10、/10、.round() 心智负担
【变更】
- app_database.dart: amount integer()() → real()()
- 所有 insert 去掉 *10 和 .round()
- 所有 display 去掉 /10
- _PointsCalculator 全部 int→double
- user_dao.dart getEarnedPoints/todayEarned 返回 double
- exit_rating_popup 配置从厘分改为分
- 服务端 IntegerField→FloatField + migration
- checkin_view reward 计算从 int(x*10) 改为直接存分
- seed_levels min_xp 厘分→分（0~481 → 0~48.1）
- build_schemas level_config REAL + 测试更新
- 设计文档 §5.4/§5.5 同步

+++ flutter_app/lib/data/daos/user_dao.dart
@@ -55,11 +55,11 @@ class UserDao {
     return result;
   }
 
-  Future<int> getEarnedPoints() async {
+  Future<double> getEarnedPoints() async {
     final rows = await (_db.select(_db.pointsTransactions)
       ..where((t) => t.source.equals('PRACTICE_REWARD'))).get();
     AuditLogger.instance.dao('UserDao.getEarnedPoints', rows.length, {});
-    var total = 0;
+    var total = 0.0;
     for (final r in rows) { total += r.amount; }
     return total;
   }
@@ -113,16 +113,16 @@ class UserDao {
   }
 
   /// 获取今天获得的学习积分（仅 earned 分类，不含 bonus/spent）
-  Future<int> getTodayEarnedPoints() async {
+  Future<double> getTodayEarnedPoints() async {
     final today = DateTime.now().toIso8601String().substring(0, 10);
     final q = (_db.select(_db.pointsTransactions)
       ..where((t) =>
           t.createdAt.isBiggerOrEqual(Variable(today)) &
-          t.amount.isBiggerThanValue(0) &
+          t.amount.isBiggerThanValue(0.0) &
           t.source.isIn(['LOGIN_BONUS', 'PRACTICE_REWARD', 'TASK_REWARD'])));
     final rows = await q.get();
     AuditLogger.instance.dao('UserDao.getTodayEarnedPoints', rows.length, {});
-    var total = 0;
+    var total = 0.0;
     for (final row in rows) {
       total += row.amount;
     }
